requirements.txt
flask==2.0.1
gunicorn==20.1.0
opencv-python-headless==4.5.3.56
numpy==1.21.2
transformers==4.11.3
torch==1.9.0
requests==2.26.0
Werkzeug==2.0.1
python-dotenv==0.19.1
# mcp>=0.1.0  # Commented out as it's not available on PyPI
Pillow>=8.0.0
